Improving SignUp UI
