+++ src/constants/modals.js
@@ -1,0 +1,58 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import {View, Modal, ActivityIndicator, Image} from 'react-native'
+import SharedStyles from '../styles/shared/sharedStyles'
+import ModalStyles from '../styles/shared/modalStyles'
+
+const styles = SharedStyles.createStyles()
+const modalStyles = ModalStyles.createStyles()
+
+export const LoadingScreen = ({toggleModal, modalVisible}) => (
+  <Modal
+    onRequestClose={() => {
+      toggleModal(!modalVisible)
+    }}
+    visible={modalVisible}
+    transparent
+  >
+    <View style={modalStyles.modalContainer}>
+      <View style={styles.flexRowCenter}>
+        <View style={modalStyles.activityIndicator}>
+          <ActivityIndicator size="large" color="#FF20B1" />
+        </View>
+      </View>
+    </View>
+  </Modal>
+)
+
+LoadingScreen.propTypes = {
+  toggleModal: PropTypes.func.isRequired,
+  modalVisible: PropTypes.bool.isRequired,
+}
+
+
+export const SuccessScreen = ({toggleModal, modalVisible}) => (
+  <Modal
+    onRequestClose={() => {
+      toggleModal(!modalVisible)
+    }}
+    visible={modalVisible}
+    transparent
+  >
+    <View style={modalStyles.modalContainer}>
+      <View style={styles.flexRowCenter}>
+        <View style={modalStyles.activityIndicator}>
+          <Image
+            style={modalStyles.emojiActivityIndicator}
+            source={require('../../assets/emoji-loader.png')}
+          />
+        </View>
+      </View>
+    </View>
+  </Modal>
+)
+
+SuccessScreen.propTypes = {
+  toggleModal: PropTypes.func.isRequired,
+  modalVisible: PropTypes.bool.isRequired,
+}
